Move stream helper method to new class. Write unit tests for stream helper.
